Updated linz so that you can now override the default Passport configuration.
