+++ lib/defaults.js
@@ -26,6 +26,7 @@ var defaults = module.exports = {
 	// error logging and output
 	'error log': console.log,
 
+	// admin ui settings
     'page size': 20,
     'page sizes': [20,50,100,200]
