+++ lib/helpers.js
@@ -1,6 +1,4 @@
 var path = require('path'),
-	passport = require('passport'),
-	LocalStrategy = require('passport-local').Strategy,
 	linz = require('../');
 
 exports.linz = function () {
@@ -18,15 +16,3 @@ exports.linz = function () {
 	};
 
 };
-
-exports.login = new LocalStrategy(function (username, password, done) {
-
-	if (username === 'deglutenous' && password === '10oct2012') {
-
-		return done(null, { username: 'deglutenous' });
-
-	}
-
-	return done(null, false, { message: 'No-op implementation.' });
-
-});
